[TASK] Avoid thread unsafe putenv() and getenv() calls
Using `putenv()` and `getenv()` is discouraged, which is can be
here https://github.com/vlucas/phpdotenv#putenv-and-getenv read
up in more detail. Main disadvantage these methods brings with
it, is the fact that they are not `threadsafe`.

Not using `getenv()` in cli contexts like for testing, provided
environment variables may be not set to $_ENV, but are contained
in the super global $_SERVER. Therefore, reading access must be
done with a `$_ENV[$key] ?? $_SERVER[$key] ?? null` chain.

* replaced `putenv()` calls with `$_ENV` array assignments
* replaced `getenv()` calls with `$_ENV` array lookups with
  proper `null coalescing operator` guarded fallback chain
  to check `$_SERVER` before returning null, but avoid php
  array key access errors

>> TYPO3 cross check against older core without adjustments
   https://review.typo3.org/c/Packages/TYPO3.CMS/+/78655

>> TYPO3 check against replacements in core
   https://review.typo3.org/c/Packages/TYPO3.CMS/+/77490

Releases: main, 7